working implementation for the HP3437A
